cleanup tests remove cli module dependency for now.
